+++ TMesh/TBot/Database/Models/PublicChannel.cs
@@ -19,7 +19,7 @@ public class PublicChannel: IRecipient
 
     byte? IRecipient.RecipientChannelXor => XorHash;
 
-    long? IRecipient.RecipientPrivateChannelId => null;
+    int? IRecipient.RecipientPrivateChannelId => null;
 
     bool? IRecipient.IsSingleDeviceChannel => null;
 }
