+++ TMesh/TBot/Models/DeviceKey.cs
@@ -23,7 +23,7 @@ namespace TBot.Models
 
         byte? IRecipient.RecipientChannelXor => null;
 
-        long? IRecipient.RecipientPrivateChannelId => null;
+        int? IRecipient.RecipientPrivateChannelId => null;
 
         bool? IRecipient.IsSingleDeviceChannel => null;
     }
